episode 24

+++ app/Http/Controllers/ThreadsController.php
@@ -111,6 +111,12 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
+        if ($thread->user_id != auth()->id()) {
+           abort(403, 'You dont have permission to do this');
+        }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
